+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add ```Request::class``` to handle http request
+- Add ```Response::class``` to handle http respone
+- Add ```RequestFactory::class``` to create request from global parameter
 
 ## [0.1.0] - 2021-07-12
 ### Added